+++ packages/admin/cms-admin/src/generator/future/generateForm.ts
@@ -255,6 +255,7 @@ export function generateForm(
         FinalFormSubmitEvent,
         Loading,
         MainContent,
+        RadioGroupField,
         TextAreaField,
         TextField,
         useFormApiRef,