+++ packages/admin/cms-admin/src/pages/pageTree/CopyPasteMenuItem.tsx
@@ -1,6 +1,5 @@
 import { RowActionsItem } from "@comet/admin";
 import { Copy, Paste } from "@comet/admin-icons";
-import React from "react";
 import { FormattedMessage } from "react-intl";
 
 import { subTreeFromNode, treeMapToArray } from "./treemap/TreeMapUtils";
@@ -12,7 +11,7 @@ interface Props {
     page: PageTreePage;
 }
 
-export const CopyPasteMenuItem = ({ page }: Props): React.ReactElement => {
+export const CopyPasteMenuItem = ({ page }: Props) => {
     const { prepareForClipboard, writeToClipboard, getFromClipboard, sendPages, progressDialog } = useCopyPastePages();
     const { tree } = usePageTreeContext();