+++ packages/api/cms-api/generate-schema.ts
@@ -29,6 +29,7 @@ import { createFolderEntity } from "./src/dam/files/entities/folder.entity";
 import { FileLicensesResolver } from "./src/dam/files/file-licenses.resolver";
 import { createFilesResolver } from "./src/dam/files/files.resolver";
 import { createFoldersResolver } from "./src/dam/files/folders.resolver";
+import { FileUploadsResolver } from "./src/file-uploads/file-uploads.resolver";
 import { RedirectInputFactory } from "./src/redirects/dto/redirect-input.factory";
 import { RedirectEntityFactory } from "./src/redirects/entities/redirect-entity.factory";
 import { AzureAiTranslatorResolver } from "./src/translation/azure-ai-translator.resolver";
@@ -111,6 +112,7 @@ async function generateSchema(): Promise<void> {
         AzureAiTranslatorResolver,
         GenerateAltTextResolver,
         GenerateImageTitleResolver,
+        FileUploadsResolver,
     ]);
 
     await writeFile("schema.gql", printSchema(schema));